Update base site styling
