+++ client/src/app/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-import Main from './Main';
-import NavBar from './NavBar';
+import Main from './main/Main';
+import NavBar from './navbar/NavBar';
+import Footer from './footer/Footer';
 import './App.css';
 
 // Base component of the app in which everything is rendered
@@ -10,6 +11,7 @@ class App extends Component {
       <div className='App'>
         <NavBar />
         <Main />
+        <Footer />
       </div>
     );
   }
