+++ client/src/app/main/Main.css
@@ -1,0 +1,20 @@
+.main {
+  width: 80%;
+  margin: 0 auto;
+  background-color: azure;
+  /* border-radius: 10px; */
+  /* box-shadow: 0px 0px 3px 3px gray; */
+}
+
+.main-background-image {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  z-index: -1;
+  background-image: url('./wood-background.jpg');
+  background-repeat: repeat;
+  background-size: contain;
+  mask-image: linear-gradient(to bottom, rgba(0, 0, 0, 1), 75%, rgba(0, 0, 0, 0));
+}
