+++ client/src/app/navbar/NavBar.css
@@ -1,0 +1,23 @@
+nav {
+  width: 100%;
+}
+
+nav ul {
+  list-style: none;
+  display: flex;
+  background-color: #01161e;
+  margin: 0;
+}
+
+nav ul li {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  padding: 10px;
+  margin: 0 10px;
+}
+
+nav ul li a {
+  color: white;
+  text-decoration: none;
+}
