--- client/src/footer/Footer.css
@@ -1,51 +1,0 @@
-.footer-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: top;
-    text-align: top;
-    background-color:#5e5e5e;
-    color:rgb(255, 255, 255);
-    min-width: 90%;
-    border-radius: 20px;
-    margin: 20px;
-  }
-
-.homepage-container h3 {
-    color: #5e5e5e;
-}
-.footer-container h3 {
-    color:antiquewhite;
-}
-
-.footer-container a {
-    color: white;
-    text-decoration: underline;
-}
-
-.footer-container a:hover {
-    color: #F56476;
-    text-decoration: underline;
-}
-
-.container {
-    min-width: 100%;
-}
-
-.homepage-container {
-    width: 80%;
-}
-
-.roundimg {
-    border-radius: 20%;
-    max-width: 80%;
-}
-
-.logo-container {
-    align-content: center;
-    padding-bottom: 30px;
-}
-.logoimg {
-    max-height: 250px;
-    padding: 20px;
-}
-
