+++ client/src/shared/Styles.css
@@ -1,35 +1,11 @@
 h1 {
-  color: #01161E;
+  color: #01161e;
 }
 
 h2 {
-  color: #01161E;
+  color: #01161e;
 }
 
 h3 {
-  color: #01161E;
-}
-
-nav {
-  width: 100%;
-}
-
-nav ul {
-  list-style: none;
-  display: flex;
-  background-color: #01161E;
-  margin: 0;
-}
-
-nav ul li {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding: 10px;
-  margin: 0 10px;
-}
-
-nav ul li a {
-  color: white;
-  text-decoration: none;
+  color: #01161e;
 }
